\name{CoCoOldData-internal}
\title{Internal CoCoOldData objects}
\alias{.clear.coco.objects}
\alias{.First.lib}
\alias{.new.coco}
\alias{.new.coco.model}
\alias{.object.of.model}
\alias{.object.of.thing}
\alias{.onLoad}
\alias{.recover.model}
\alias{.recover.reference}
\alias{.return.key}
\alias{.return.model.number}
\alias{.return.model.of.object}
\alias{.return.object.model.number}
\alias{.return.reference}
\alias{.return.type}
\alias{.set.asymptotic}
\alias{.set.coco.value}
\alias{.set.em.epsilon}
\alias{.set.em.initial}
\alias{.set.em.max.iterations}
\alias{.set.exact.epsilon}
\alias{.set.exact.test}
\alias{.set.ips.epsilon}
\alias{.set.ips.max.iterations}
\alias{.set.ips.stop.criterion}
\alias{.set.list.of.number.of.tables}
\alias{.set.number.of.tables}
\alias{.set.page.formats}
\alias{.set.paging.length}
\alias{.set.power.lambda}
\alias{.set.print.formats}
\alias{.set.seed.coco}
\alias{.SetSlotValue}
\alias{.set.table.formats}
\alias{.set.test}
\alias{.set.test.formats}

\alias{label} % CoCoTestClass-method
\alias{testLabel} % CoCoTestClass-method
\alias{width} % CoCoTestClass-method
\alias{testWidth} % CoCoTestClass-method

\alias{dynamic.Graph} % CoCoModelClass-method
\alias{dynamic.Graph.Old} % CoCoModelClass-method
\alias{modifyModel} % CoCoModelClass-method
\alias{subModifyModel} % CoCoModelClass-method
\alias{subModifyModel.Old} % CoCoModelClass-method
\alias{testEdge} % CoCoModelClass-method
\alias{testEdge.Old} % CoCoModelClass-method

\alias{makeModel} % CoCoObjects
\alias{newCoCoTestObject} % CoCoObjects
\alias{recover.coco.model} % CoCoObjects

\alias{coco.init} % From CoCoCore
\alias{coco.resume} % From CoCoCore
\alias{coco.start} % From CoCoCore
\alias{currentCoCo} % From CoCoObjects
\alias{endCoCo} % From CoCoRaw
\alias{is.number} % From CoCoCore
\alias{make.cococg} % From CoCoCore
\alias{make.coco} % From CoCoCore
\alias{makeCurrentCoCo} % From CoCoObjects
\alias{make.model} % From CoCoCore
\alias{my.assign} % From CoCoCore
\alias{quitCoCo} % From CoCoRaw

\description{Internal CoCoOldData and imported objects.}
\details{These are not to be called by the user.}
\keyword{internal}
