\name{CoCoOldData-package}
\alias{CoCoOldData-package}
\docType{package}
\title{CoCoOldData}
\description{
This package contains functions for reading ASCII files 
with data for CoCo in a deprecated format.

These are not to be called by the user.

Please use the functions of the package \link[CoCoRaw]{CoCoRaw-package}
(and \link[CoCoObjects]{CoCoObjects-package}
 and \link[CoCoGraph]{CoCoGraph-package}). 

Datasets are found in the packages \link[CoCo]{CoCo-package} and
\link[CoCoCg]{CoCoCg-package}.
}
\details{
   \tabular{ll}{
      Package: \tab CoCoOldData  \cr
      Type:    \tab Package      \cr
      Version: \tab 0.1.7.5      \cr
      Date:    \tab 2009-04-03   \cr
   }
}
\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
